% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkAddChemGadget.R
\name{addChemsToHTTK}
\alias{addChemsToHTTK}
\title{Function that adds chemical to httk package chem list.}
\usage{
addChemsToHTTK()
}
\description{
internal function that interfaces with httk package
}
\examples{
\dontrun{
addChemsToHTTK()
}
}
