% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDatasetModule.R
\name{addDataSet}
\alias{addDataSet}
\title{Server side function for the UI used to add external observation datasets to projects}
\usage{
addDataSet(input, output, session, data_type)
}
\arguments{
\item{input}{the input object from the add dataset module UI}

\item{output}{the output object from the add dataset module UI}

\item{session}{the shiny session information where the add dataset in called currently}

\item{data_type}{the data_type returned by the module.}
}
\description{
The sever function for the add dataset module used in PLETHEM. This function interacts with the server function
  of the PLETHEM model. It saves the dataset imported as a .RDS file in the project folder. It also adds the dataset reference
  to the project database for further use.
}
