% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDatasetModule.R
\name{addDataSetUI}
\alias{addDataSetUI}
\title{UI for adding external observation datasets to projects}
\usage{
addDataSetUI(namespace, data_type)
}
\arguments{
\item{namespace}{the namespace for this module}

\item{data_type}{the type of data to be uploaded based on where the UI is called from}
}
\description{
The UI function for the add dataset module used in PLETHEM. This function creates the add dataset dialog
  box that is triggered by the modeling interface. This is a module function and the user will never need to call it
}
