% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{addParameters}
\alias{addParameters}
\title{Add user selected parameters to the PBPK reporting document.
Should not be used by directly by the user}
\usage{
addParameters(HESI_doc, chemical, exposure, physiological)
}
\arguments{
\item{HESI_doc}{This is the PBPK modeling document that the flowchart will be added too.}

\item{chemical}{dataframe of chemical data added by the user}

\item{exposure}{dataframe of exposure data added by the user}

\item{physiological}{dataframe of physiological data added by the user}
}
\value{
HESI_doc, the same officer::read_docx object
}
\description{
add user defined parameter tables
}
