% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{addPBPKequations}
\alias{addPBPKequations}
\title{Add PBPK equations to the PBPK reporting document. Should not be used by directly by the user}
\usage{
addPBPKequations(HESI_doc)
}
\arguments{
\item{HESI_doc}{This is the PBPK modeling document that the flowchart will be added too.}
}
\value{
PBPK modeling document with equations and starting parameter values.
}
\description{
The function returns the PBPK reporting document with the PBPK modeling equations
and starting variable values to it.
}
