% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomDataExample.R
\docType{data}
\name{biomDataExample}
\alias{biomDataExample}
\title{NHANES Toluene Bio-monitoring Data Set For Testing The runReverseDosimetry Function}
\format{
An object of class \code{data.frame} with 10000 rows and 1 columns.
}
\source{
\href{https://www.scitovation.com/plethem/}{PLETHEM}
}
\usage{
data(biomDataExample)
}
\description{
This dataset was taken from the Centers for Disease Control, National Health
and Nutritional Examination Survey data. It has been significantly modified
in its formatting and only contains data on toluene biomonitering for testing
the runReverseDosimetry Function. The original data is availible through the
NHANES reference. The modified data set is availible through the source
reference.
}
\examples{
data(biomDataExample)
}
\references{
\href{https://www.cdc.gov/nchs/nhanes/index.htm}{NHANES}
}
\keyword{NHANES}
\keyword{Toluene}
