% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{bloodConcRange}
\alias{bloodConcRange}
\title{Function to create reverse dosimetry headers}
\usage{
bloodConcRange(minValue = -0.6, maxValue = 4.6, stepSize = 0.2)
}
\value{
upperAndLowerBounds
}
\description{
ReverseDosimetryDataHeaders.csv
}
