% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{bloodConcRange2}
\alias{bloodConcRange2}
\title{Function to create reverse dosimetry headers}
\usage{
bloodConcRange2(mcDF = data.frame())
}
\value{
upperAndLowerBounds
}
\description{
ReverseDosimetryDataHeaders.csv
}
