% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HepaticMetabolism.R
\name{calc_cyp2b6}
\alias{calc_cyp2b6}
\title{Calculate cyp2b6 fraction}
\usage{
calc_cyp2b6(age)
}
\arguments{
\item{age}{age of the human in years}
}
\description{
get expression for cyp2b6 as a fraction of adult
}
