% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HepaticMetabolism.R
\name{calcMPCPPGL}
\alias{calcMPCPPGL}
\title{Calcuate MPPGL and CPPGL based on age}
\usage{
calcMPCPPGL(age)
}
\arguments{
\item{age}{age of the human in years}
}
\value{
list containing the "MPPGL" and "CPPGL" values for the
}
\description{
Calculates the Microsomal Protein per Gram liver (MPPGL) and Cytosolic Protein Per Gram Liver(CPPGL)
  for humans based on the age of the person in years. The equations for MPPGL and CPPGL were developed internally at
  Scitovation
}
