% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{calcPercentileValue}
\alias{calcPercentileValue}
\title{Calculate percentile values for exposure}
\usage{
calcPercentileValue(whichPercentile = 25, cdf, exposure_list)
}
\description{
Calulates the percentile value for a specific percentile when provided with the CDF and list of exposures for the CDF.
}
