% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateInhVolCss}
\alias{calculateInhVolCss}
\title{calculate steady state concentration of inhaled volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence}
\usage{
calculateInhVolCss(dose, ql, qc, qp, pair, scaled_hepcl, scaled_plcl)
}
\description{
calculate steady state concentration of inhaled volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence
}
