% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateOralNonvolCss}
\alias{calculateOralNonvolCss}
\title{calculate steady state concentration of orally adminitered non volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence}
\usage{
calculateOralNonvolCss(dose, bw, ql, qc, fup, cl_type, scaled_hepcl,
  scaled_rencl, scaled_plcl)
}
\description{
calculate steady state concentration of orally adminitered non volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence
}
