% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateOralVolCss}
\alias{calculateOralVolCss}
\title{calculate steady state concentration of orally administered volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence}
\usage{
calculateOralVolCss(dose, bw, ql, qc, pair, scaled_hepcl, scaled_plcl)
}
\description{
calculate steady state concentration of orally administered volatile compounds
based on steady state clearence for renal,#'hepatic and plasma clearence
}
