% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateRecombClearance}
\alias{calculateRecombClearance}
\title{Calculate scaled in-vivo clearance if in-vitro clearance is meausured using S9 Fraction}
\usage{
calculateRecombClearance(clearance, organism, age, liver_wt, cyp_data,
  return_total = T)
}
\arguments{
\item{clearance}{number for measured S9 clearance}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}

\item{cyp_data}{dataframe containing all the cyp values with column names (name,loc,abundance,isef,fumic)}

\item{return_total}{logical, type of value to be returned,}
}
\value{
Total scaled invivo clearance in L/h if return_total is TRUE, else a named list of individual clearances
}
\description{
In-vitro clearance is measured in S9 fraction. The function uses measured in-vitro clearance
and physiological parameters to estimate in-vivo clearance. This function is not called directly by the user
}
