% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateRecombClearence}
\alias{calculateRecombClearence}
\title{Calculate scaled in-vivo clearence if in-vitro clearence is meausured using S9 Fraction}
\usage{
calculateRecombClearence(clearence, organism, age, liver_wt, cyp_data,
  return_total = T)
}
\arguments{
\item{clearence}{number for measured S9 clearence}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}

\item{cyp_data}{dataframe containing all the cyp values with column names (name,loc,abundance,isef,fumic)}

\item{return_total}{logical, type of value to be returned,}
}
\value{
Total scaled invivo clearence in L/h if return_total is TRUE, else a named list of individual clearances
}
\description{
In-vitro clearence is measured in S9 fraction. The function uses measured in-vitro clearence
and physiological parameters to estimate in-vivo clearence. This function is not called directly by the user
}
