% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateScaledS9Clearance}
\alias{calculateScaledS9Clearance}
\title{Calculate scaled in-vivo clearance if in-vitro clearance is meausured using S9 Fraction}
\usage{
calculateScaledS9Clearance(clearance, units, organism, age, liver_wt,
  km = 1, mpcppgl = list())
}
\arguments{
\item{clearance}{number for measured S9 clearance}

\item{units}{string for S9 clearance units}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}
}
\value{
Scaled invivo clearance in L/h
}
\description{
In-vitro clearance is measured in S9 fraction. The function uses measured in-vitro clearance
and physiological parameters to estimate in-vivo clearance. This function is not called directly by the user
}
