% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateScaledS9Clearence}
\alias{calculateScaledS9Clearence}
\title{Calculate scaled in-vivo clearence if in-vitro clearence is meausured using S9 Fraction}
\usage{
calculateScaledS9Clearence(clearence, units, organism, age, liver_wt)
}
\arguments{
\item{clearence}{number for measured S9 clearence}

\item{units}{string for S9 clearence units}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}
}
\value{
Scaled invivo clearence in L/h
}
\description{
In-vitro clearence is measured in S9 fraction. The function uses measured in-vitro clearence
and physiological parameters to estimate in-vivo clearence. This function is not called directly by the user
}
