% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateScaledSCClearance}
\alias{calculateScaledSCClearance}
\title{Calculate scaled in-vivo clearance if in-vitro clearance is meausured using Sub-cellular components}
\usage{
calculateScaledSCClearance(clearance, units, organism = "human", age,
  liver_wt, km, mpcppgl = list(), return_total = T)
}
\arguments{
\item{clearance}{vector of (microsomal,cytosolic) clearance}

\item{units}{vector of (microsomal,cytosolic) clearance units}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}

\item{return_total}{logical, type of value to be returned,}
}
\value{
Total scaled invivo clearance in L/h if return_total is TRUE, else a named list of individual clearances
}
\description{
In-vitro clearance is measured in microsomal and cytosolic fractions. The function uses measured in-vitro clearance
and physiological parameters to estimate in-vivo clearance. This function is not called directly by the user
}
