% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateScaledSCClearence}
\alias{calculateScaledSCClearence}
\title{Calculate scaled in-vivo clearence if in-vitro clearence is meausured using Sub-cellular components}
\usage{
calculateScaledSCClearence(clearence, units, organism = "human", age,
  liver_wt, km, return_total = T)
}
\arguments{
\item{clearence}{vector of (microsomal,cytosolic) clearence}

\item{units}{vector of (microsomal,cytosolic) clearence units}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}

\item{liver_wt}{liver weight(kg)}

\item{return_total}{logical, type of value to be returned,}
}
\value{
Total scaled invivo clearence in L/h if return_total is TRUE, else a named list of individual clearances
}
\description{
In-vitro clearence is measured in microsomal and cytosolic fractions. The function uses measured in-vitro clearence
and physiological parameters to estimate in-vivo clearence. This function is not called directly by the user
}
