% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{calculateScaledWholeHepClearance}
\alias{calculateScaledWholeHepClearance}
\title{Calculate sclaed in-vivo clearance if in-vitro clearance is measured in whole hepatocytes}
\usage{
calculateScaledWholeHepClearance(clearance, units, liver_wt, hpgl, km = 1)
}
\arguments{
\item{clearance}{number for measured whole hepatocyte clearance}

\item{units}{string for whole hepatocyte clearance units}

\item{liver_wt}{liver weight(kg),}

\item{km}{michelis-menten constant for chemical}

\item{organism}{name of organism, "human" or "rat"}

\item{age}{age of organism in "years" if "human" or "weeks" if "rat"}
}
\value{
Scaled invivo clearance in L/h
}
\description{
The functionuses measured in-vitro clearance in whole hepatocytes, and phyiological parameters, 
to estimate in-vivo clearance. This function is not called directly by the user.
}
