% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{cdf}
\alias{cdf}
\title{calcualte CDF from the probability table}
\usage{
cdf(probabilityTable)
}
\description{
Converts a PDF to a CDF so it can be used for plotting and calculating percentile values
}
