% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalDbModule.R
\name{chemical}
\alias{chemical}
\title{Server side function for the UI used to edit local chemical database}
\usage{
chemical(input, output, sessions)
}
\arguments{
\item{input}{the input object from the chemical module UI}

\item{output}{the output object from the chemical module UI}

\item{session}{the shiny session information where the chemical module is currently called}
}
\description{
The sever function for the chemical module used in PLETHEM. This function interacts with the server function
  of the PLETHEM model.Is modifies the local chemical database by adding, deleting or modifying existing chemical data.
}
