% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalDbModule.R
\name{chemicalInput}
\alias{chemicalInput}
\title{UI for modufying the user chemical database}
\usage{
chemicalInput(namespace)
}
\arguments{
\item{namespace}{the namespace for this module}
}
\description{
The UI function for calling the module that allows the user to modify the local chemical database.
  This is a module function and the user will never need to call it
}
