% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{clearProjectDb}
\alias{clearProjectDb}
\title{Clear Project Db}
\usage{
clearProjectDb()
}
\description{
This function clears the project Db. It is called internally when a new project is created.
It is also used by developers to make a clean project db
}
