% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDb.R
\name{createEmptyUserDb}
\alias{createEmptyUserDb}
\title{Creates a new userDb based on the empty database in the package}
\usage{
createEmptyUserDb(path = NULL)
}
\arguments{
\item{path}{path to where the user database needs to be stored. Make sure you have write permission to this folder. If no path is provided, it launches a folder select dialog.}
}
\description{
The function allows the users to create a new empty user database file that is needed to run plethem.
}
\examples{
\dontrun{
createEmptyUserDb(),
createEmptyUserDb("C:/Users/Documents/")
}
}
