% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{createHESIgraphs}
\alias{createHESIgraphs}
\title{Pipe to add concentration-timecourse data to the report
Should not be used by directly by the user}
\usage{
createHESIgraphs(report_doc, context, conc_units)
}
\arguments{
\item{report_doc}{an officer::read_docx object}

\item{context}{a data.frame of concentrationXtime data, as returned by concData()}

\item{conc_units}{a string denoting the units of concentration}
}
\value{
report_doc, the same officer::read_docx object
}
\description{
add graphs to the PBPK reporting document
}
