% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{createPBPKflowchart}
\alias{createPBPKflowchart}
\title{Creates the PBPK model flowchart. Should not be used by directly by the user}
\usage{
createPBPKflowchart(HESI_doc)
}
\arguments{
\item{HESI_doc}{This is the PBPK modeling document that the flowchart will be added too.}
}
\value{
The PBPK modeling document with the PBPK flowchart added to it.
}
\description{
The function returns the PBPK model flowchart
}
