% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDb.R
\name{createUserDb}
\alias{createUserDb}
\title{Creates a new userDb based on the empty database in the package}
\usage{
createUserDb()
}
\description{
Creates a new userDb based on the empty database in the package
}
