% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIVIVE.R
\name{doIVIVE}
\alias{doIVIVE}
\title{run IVIVE gadget and save files as needed.}
\usage{
doIVIVE()
}
\value{
Values returned by the IVIVE gadget
}
\description{
Run IVIVE gadget and save CSV files needed to create metabolism sets in the PBPK model
}
\examples{
\dontrun{
doIVIVE()
}
}
