% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editUserDbGadget.R
\name{editUserDb}
\alias{editUserDb}
\title{Gadget for creating/editing the a user database}
\usage{
editUserDb(file_path = "")
}
\arguments{
\item{file_path}{Path of the database where changes are to be made}
}
\description{
This function defines the UI and server for creating a new user database. It accepts CSV files from users for chemicals and exposure.
}
