% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModels.R
\name{executeInteractivePBPKWorkflows}
\alias{executeInteractivePBPKWorkflows}
\title{Run PBPK workflows}
\usage{
executeInteractivePBPKWorkflows(
  model = "rapidPBPK",
  simid = NULL,
  progressFunc = NULL
)
}
\arguments{
\item{model}{The name of the PBPK model to simulate. Defaults to 'rapidPBPK'.}

\item{simid}{Simulation ID for the simulation to run from the project database in PLETHEM.
This function should not be called by the user from the console
Defaults to 'fd'.}
}
\description{
Run all PBPK worfklows from the plethem package.
This function is called by the User interface to run the selected simulation.
See\code{\link{runFDPBPK}} for running forward dosimetry models from the console.
}
