% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureTRAGadget.R
\name{exposureTRAGadget}
\alias{exposureTRAGadget}
\title{Gadget UI and server for importing data run using Commercial TRA version}
\usage{
exposureTRAGadget(save_flag = F, base_path = NULL)
}
\description{
Gadget UI and server for importing data run using Commercial TRA version
}
