% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{externDbSelect}
\alias{externDbSelect}
\title{Runs all select queries an arbitrary database}
\usage{
externDbSelect(query, db_path)
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{A valid path}
}
\description{
The function runs the select queries issued to the user db and returns the dataframe
the path to user database is stored in main plethem database and is selected from there
}
