% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HepaticMetabolism.R
\name{getAllCypData}
\alias{getAllCypData}
\title{get data for human cyp fraction by age}
\usage{
getAllCypData(age)
}
\arguments{
\item{age}{the age of the person in years}
}
\value{
dataframe containing the fractional expression for all the cyps
}
\description{
The function returns the expression of cyps at each age as a fraction of the adult (25 years).
  The number needs to be multiplied by the cyp abundance information in the database to get the nominal expression
  value for the cyp. See \code{\link{getCypAbundance}} on how to get abundance values for a given age
}
