% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getAllParamNames}
\alias{getAllParamNames}
\title{Gets all the param_names for the model}
\usage{
getAllParamNames(model)
}
\arguments{
\item{model}{Character string indentfying the model to be run}
}
\description{
Get all the param_names for any given model. The function is not called by the user
}
