% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getAllParamValuesForModel}
\alias{getAllParamValuesForModel}
\title{Gets all the parameter values for the model. This function should not be used by the model}
\usage{
getAllParamValuesForModel(simid, model)
}
\arguments{
\item{simid}{Integer The id for simulation selected to run}

\item{model}{Character The string identifying the model to be run}
}
\value{
list List that can be passed to the solver as model params
}
\description{
Get all the parameter values that are required for the model to run. The values
are obtained from the Project database. Only those values that are used in the model as
determined by the master database are returned by the function
}
