% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getAllSetChoices}
\alias{getAllSetChoices}
\title{get all set names for a given paramter set}
\usage{
getAllSetChoices(set_type = "physio")
}
\arguments{
\item{set_type}{The type of set can be "physio", "chem","expo","metab" or "sim"}
}
\value{
named list of set names
}
\description{
This function returns all the sets of a given set type from the current project database
This is used internally to update drop downs or to get simulation choices
}
