% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{getChemDataFromDb}
\alias{getChemDataFromDb}
\title{Get the chemical data from the project Db
This function is not designed to be used by the end user}
\usage{
getChemDataFromDb(chemid = NULL)
}
\description{
Get the chemical data from the project Db
This function is not designed to be used by the end user
}
