% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{getDataFrequency}
\alias{getDataFrequency}
\title{This function bins the data based and calcualtes frequency of data in each bin}
\usage{
getDataFrequency(bounds, mcResult)
}
\description{
This function bins the data based and calcualtes frequency of data in each bin
}
