% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{getDataProbability}
\alias{getDataProbability}
\title{Function to calculate probability from the frequency table}
\usage{
getDataProbability(frequencyTable, weightingFactor = c())
}
\description{
The function calculates probability from the frequency table. If weighting factor is empty, 
the frequency table is assumed to be counts that are to be normalized. If a weighting factor array is provided,
the frequency table is assumed to be a fraction that needs to be scaled by the weighting factor.
}
