% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{getFileFolderPath}
\alias{getFileFolderPath}
\title{Show dialogs to select files or folders}
\usage{
getFileFolderPath(type = "dir", caption, extension)
}
\description{
The function shows the dialog to select files or folders. The functions change depending on the OS in which 
RStudio is running. It is only called internally
}
