% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseBodyHeight}
\alias{getLifecourseBodyHeight}
\title{Get average body height using the life course equation}
\usage{
getLifecourseBodyHeight(age, gender)
}
\arguments{
\item{age}{age in years}

\item{gender}{Either "M" for male or "F" for female}
}
\description{
The function is used to calculate the average body height for humans based on age and gender.
 The function uses life course equations developed by Scitovation.
}
