% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseBodyHeightMale}
\alias{getLifecourseBodyHeightMale}
\title{Calculate body height of average human male using the lifecourse
equation}
\usage{
getLifecourseBodyHeightMale(age)
}
\arguments{
\item{age}{age in years}
}
\value{
Body height
}
\description{
This function is not called directly by the user. See
  \code{\link{getLifeCourseBodyHeight}}
}
