% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseBodyWeight}
\alias{getLifecourseBodyWeight}
\title{Get average body weight using the life course equation}
\usage{
getLifecourseBodyWeight(age, gender)
}
\arguments{
\item{age}{age in years}

\item{gender}{Either "M" for male or "F" for female}
}
\value{
Body Weight in kgs
}
\description{
The function is used to calculate the average body weight in kgs for humans based on age and gender.
 The function uses life course equations developed by Scitovation.
}
