% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseBodyWeightMale}
\alias{getLifecourseBodyWeightMale}
\title{Calculate body weight in kg of average human male using the lifecourse
equation}
\usage{
getLifecourseBodyWeightMale(age)
}
\arguments{
\item{age}{age in years}
}
\value{
Body weight in kg
}
\description{
This function is not called directly by the user. See
  \code{getLifeCourseBodyWeight}
}
