% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseCardiacOutput}
\alias{getLifecourseCardiacOutput}
\title{Get the cardiac output based on the life course equations}
\usage{
getLifecourseCardiacOutput(age, gender, source = "sciv", qc_var = 0)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Gender can be either "M" for male or "F" for female}

\item{source}{Source of the equations. Either from Bosgra et al. "bosgra" or from ScitoVation publications as "sciv"}

\item{qc_var}{variability for cardiac output. Used to account for exercise level. Can only be used when source = "sciv"}
}
\description{
Get the cardiac output for a specific age and gender. Uses either Bosgra or ScitoVation equations
}
