% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseGlomerularFiltrationRate}
\alias{getLifecourseGlomerularFiltrationRate}
\title{Get the Glomerular Filteration Rate for a given age and gender}
\usage{
getLifecourseGlomerularFiltrationRate(age, gender)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Either "M" for Male for "F" for Female}
}
\description{
Uses life course equation to calculate a GFR for a specific age and gender
}
