% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseLungDeadSpace}
\alias{getLifecourseLungDeadSpace}
\title{Get the Lung Dead Space for a given age and gender}
\usage{
getLifecourseLungDeadSpace(age, gender)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Either "M" for Male for "F" for Female}
}
\description{
Uses life course equation to calculate the lung dead space for a specific age and gender
}
