% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseTidalVolume}
\alias{getLifecourseTidalVolume}
\title{Get the tidal volume for a given age and gender}
\usage{
getLifecourseTidalVolume(age, gender, activity = "rest")
}
\arguments{
\item{age}{Age in years}

\item{gender}{Either "M" for Male for "F" for Female}

\item{activity}{Activity level. Can be "rest" or "light activity"}
}
\description{
Uses life course equation to calculate the tidal volume for a specific age and gender
}
