% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseTissuePerfusion}
\alias{getLifecourseTissuePerfusion}
\title{Get perfusion for tissues provided on the basis of age and gender}
\usage{
getLifecourseTissuePerfusion(age = 25, gender = "M", tissues = list(),
  source = "sciv")
}
\arguments{
\item{age}{Age of the organism in years}

\item{gender}{The gender of the organism "M" for Male or "F" for Female}

\item{tissues}{List of tissues for which blood perfusion is needed. The
tissues can be any of the following
liver,
fat,
bone,
brain,
gonad,
heart,
intestine,
kidney,
lung,
pancreas,
skin,
spleen,
stomach,
thymus,
muscle,
remaining}
}
\value{
list containing blood perfusion of tissues for the given age and
gender.
}
\description{
Get perfusion for tissues provided on the basis of age and gender
}
