% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseTissueVolumes}
\alias{getLifecourseTissueVolumes}
\title{Get volumes for tissues provided on the basis of age and gender}
\usage{
getLifecourseTissueVolumes(age = 25, gender = "M", nonperf_frct = 0.85,
  tissues = list())
}
\arguments{
\item{age}{Age of the organism in years}

\item{gender}{The gender of the organism "M" for Male or "F" for Female}

\item{tissues}{List of tissues for which the volumes are needed. The tissues can be one of the following
liver,
blood,
fat,
bone,
brain,
gonad,
heart,
intestine,
kidney,
lung,
pancreas,
skin,
spleen,
stomach,
thymus,
remaining,
muscle}
}
\value{
list containing volumes for age, gender and tissues.
}
\description{
Get volumes for tissues provided on the basis of age and gender
}
