% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseUrineProductionRate}
\alias{getLifecourseUrineProductionRate}
\title{Get the urine production rate for a given age and gender}
\usage{
getLifecourseUrineProductionRate(age, gender)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Either "M" for Male for "F" for Female}
}
\description{
Uses life course equation to calculate a urine production rate for a specific age and gender
}
