% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{getLifecourseVentilationRate}
\alias{getLifecourseVentilationRate}
\title{Get the ventilation rate for a given age and gender}
\usage{
getLifecourseVentilationRate(age, gender, activity = "rest", source = "sciv")
}
\arguments{
\item{age}{Age in years}

\item{gender}{Either "M" for Male for "F" for Female}

\item{activity}{Activity level. Can be "rest" or "light exercise"}

\item{source}{source of equations either "sciv" or "bosgra"}
}
\description{
Uses life course equation to calculate a ventilation rate for a specific age and gender
}
