% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{getMCConcBins}
\alias{getMCConcBins}
\title{Create bins for concentrations from the monte carlo data}
\usage{
getMCConcBins(mcDF = data.frame())
}
\description{
Create bins for concentrations from the monte carlo data
}
