% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getMetabData}
\alias{getMetabData}
\title{Gets the metabolism data. Should not be used by directly by the user}
\usage{
getMetabData(admeid, model = "rapidPBPK")
}
\arguments{
\item{admeid}{The id for ADME set. The admeid is used to obtain information about the other sets.}

\item{model}{Model name}
}
\value{
List containing the metabolism values needed to run PBPK model or
display simulation information
}
\description{
The function returns the relavent metabolism data if the simulation contains
data from the metabolism set
}
