% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{getNextID}
\alias{getNextID}
\title{Get the next Id for the table entry}
\usage{
getNextID(tble_name, db_path = "database/project.sqlite")
}
\arguments{
\item{tble_name}{Name of the table for which the ID is required}

\item{db_path}{The location of the project databse.
This defualts to database/project.sqlite and is not expected to change}
}
\value{
int Integer ID of the next row in the table.
}
\description{
Gets the next valid id for a given table.
This function is used internally to decide what id number should be used to save a parameter set
}
