% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{getObsDataDistribution}
\alias{getObsDataDistribution}
\title{Function to generate histogram of data}
\usage{
getObsDataDistribution(obsData, concBins)
}
\description{
The function returns a distribution for the observed data by using 
the bins generated using the monte carlo simulation data
}
