% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getObservationSetChoices}
\alias{getObservationSetChoices}
\title{Get all observation sets}
\usage{
getObservationSetChoices(obs_type)
}
\arguments{
\item{obs_type}{type of observation to return, can be "cl" or "conc" for clearance and concentration data}
}
\value{
named list of all sets of the obs_type
}
\description{
Get all the sets associated with observation in a given projects. Observations need to
be handled differently from the other set types since they can themselves be of multiple types
}
