% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getParameterSet}
\alias{getParameterSet}
\title{Get the values for parameters in a given set}
\usage{
getParameterSet(set_type = "physio", id = 1)
}
\arguments{
\item{set_type}{Either "physio","chem"or "expo"}

\item{id}{integer id for the required set}
}
\description{
Get all the parameter values for a given dataset and id
}
