% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getVariabilitySetChoices}
\alias{getVariabilitySetChoices}
\title{Get all variability sets}
\usage{
getVariabilitySetChoices(var_type = "physio")
}
\arguments{
\item{var_type}{type of Variability set to return, can be "physio","chem" or "expo" or "conc"}
}
\value{
named list of all sets of the var_type
}
\description{
Get all the variability datasets in a given projects. Variabilities need to
be handled differently from the other set types since they can themselves be of multiple types
}
