% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE_InputModule.R
\name{HT_IVIVEUI}
\alias{HT_IVIVEUI}
\title{Module for editing high throughput reverse dosimetry functions}
\usage{
HT_IVIVEUI(namespace = "")
}
\description{
Module for editing high throughput reverse dosimetry functions
}
