% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkOralEqDoseGadget.R
\name{httkCalcOralEqDose}
\alias{httkCalcOralEqDose}
\title{Function that runs the httk oral equivalent dose gadaget.}
\usage{
httkCalcOralEqDose()
}
\description{
Function that runs the httk oral equivalent dose gadaget.
}
