% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkUtils.R
\name{httkParameterPBTK}
\alias{httkParameterPBTK}
\title{Call httk function for parameterizing PBPK models}
\usage{
httkParameterPBTK(chem_name, species = "Human")
}
\arguments{
\item{chem_name}{name of the chemical for which to parameterize the model. Has to be in the httk database.}

\item{species}{species for which to parameteize the model for. Defaults to human}
}
\value{
list containing parameters for the PBPK model
}
\description{
Function that calls parameterize_pbpk function within the httk package
}
