% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllExposureModule.R
\name{importAllExposureData}
\alias{importAllExposureData}
\title{Server function for seem data module}
\usage{
importAllExposureData(input, output, session, expo_name_df)
}
\arguments{
\item{input}{input object for the UI}

\item{output}{input object to the UI}

\item{session}{session object for the module}

\item{expo_name_df}{dataframe containing variable names for exposure values}

\item{fpath_seems}{path to the SEEM database}
}
\description{
Server function for import seem data module. This function should not be called by the user
}
