% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllExposureModule.R
\name{importAllExposureDataUI}
\alias{importAllExposureDataUI}
\title{UI for importing SEEM data.}
\usage{
importAllExposureDataUI(namespace)
}
\arguments{
\item{namespace}{namespace for the module UI.}
}
\description{
This function is called by the pbpk model to import SEEM exposure estimates. Never called by the user
}
