% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBatchExposureModule.R
\name{importBatchExposure}
\alias{importBatchExposure}
\title{Server function for batch data module}
\usage{
importBatchExposure(input, output, session, expo_name_df)
}
\arguments{
\item{input}{input object for the UI}

\item{output}{input object to the UI}

\item{session}{session object for the module}

\item{expo_name_df}{dataframe containing variable names for exposure values}
}
\description{
Server function for import batch data module. This function should not be called by the user
}
