% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBatchExposureModule.R
\name{importBatchExposureUI}
\alias{importBatchExposureUI}
\title{UI function for importing generic exposure data}
\usage{
importBatchExposureUI(namespace)
}
\arguments{
\item{namespace}{namespace for the module}
}
\description{
UI function for importing generic exposure data into PLETHEM. The function should not be called by the user
}
