% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHTTKDataModule.R
\name{importHTTKData}
\alias{importHTTKData}
\title{Server function for importing HTTK data}
\usage{
importHTTKData(input, output, session)
}
\description{
This is the server function for the importHTTKData module. This is not to be called by the user.
}
