% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importParameterSetModule.R
\name{importParameterSet}
\alias{importParameterSet}
\title{Server for import parameter module}
\usage{
importParameterSet(input, output, session, set_type)
}
\description{
Server for import parameter module
}
