% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importParameterSetModule.R
\name{importParameterSetUI}
\alias{importParameterSetUI}
\title{UI for importing parameter sets}
\usage{
importParameterSetUI(namespace, set_type)
}
\description{
UI for importing parameter sets
}
