% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSHEDSDataModule.R
\name{importShedsData}
\alias{importShedsData}
\title{Server function for importing SHEDS data}
\usage{
importShedsData(input, output, session, path, expo_name_df)
}
\description{
Server function for importing SHEDS exposures estimates into PLETHEM. The function should not be called by the user
}
