% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSHEDSDataModule.R
\name{importShedsDataUI}
\alias{importShedsDataUI}
\title{UI function for importing SHEDS data}
\usage{
importShedsDataUI(namespace)
}
\description{
UI function for importing SHEDS exposures estimates into PLETHEM. The function should not be called by the user
}
