% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactiveHT}
\alias{interactiveHT}
\title{Launch HT-IVIVE interface}
\usage{
interactiveHT(name = "HT-IVIVE")
}
\arguments{
\item{name}{name of the  model. Has to be "HT-IVIVE"}
}
\description{
Used internally to launch the HT-IVIVE UI. HT-IVIVE does not use the project management system that PBPK models uses.
}
\examples{
if(interactive()){
interactiveHT("HT-IVIVE")
}
}
