% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactivePBPK}
\alias{interactivePBPK}
\title{Launch PBPK interface
valid values for names are "rapidPBPK"}
\usage{
interactivePBPK(name = "")
}
\description{
Launch PBPK interface
valid values for names are "rapidPBPK"
}
