% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactiveReverseDosimetry}
\alias{interactiveReverseDosimetry}
\title{Launch Reverse Dosimetry Interface}
\usage{
interactiveReverseDosimetry()
}
\description{
Used to launch the reverse dosimetry UI. The UI is identical to the one found on \href{https://scitovation.shinyapps.io/ReverseDosimetry/}{ShinyApps Reverse Dosimetry App}
}
\examples{
\dontrun{
interactiveReverseDosimetry()
} 
}
