% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactiveRouteToRoute}
\alias{interactiveRouteToRoute}
\title{Launch RouteToRouteExtrapolation Interface}
\usage{
interactiveRouteToRoute()
}
\description{
Used to launch the route to route UI. The UI is identical to the one found on \href{https://scitovation.shinyapps.io/ReverseDosimetry/}{ShinyApps Reverse Dosimetry App}
}
\examples{
\dontrun{
interactiveRouteToRoute()
} 
}
