% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifecoursePhysiology.R
\name{interp2}
\alias{interp2}
\title{Small utility function for linear interpolation
Used internally}
\usage{
interp2(x1, x2, y1, y2, x)
}
\description{
Small utility function for linear interpolation
Used internally
}
