% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{isBetween}
\alias{isBetween}
\title{bolean to check if a number is within a given range}
\usage{
isBetween(largerThan, smallerThanEqualTo, testValue)
}
\description{
bolean to check if a number is within a given range
}
