% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEGadget.R
\name{iviveGadget}
\alias{iviveGadget}
\title{Command line function for running the IVIVE gadget}
\usage{
iviveGadget(save_flag = F, base_path = NULL)
}
\arguments{
\item{save_flag}{Boolean. Should the data generated be saved to a csv file. The name of the file can then be defined in the UI}

\item{base_path}{if save_flag is True, must specify a base path where the result file will be}
}
\description{
Launches the IVIVE gadget for estimating age dependent metabolism
}
