% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{listProjects}
\alias{listProjects}
\title{List all the project currently in the PLETHEM}
\usage{
listProjects(type = NULL)
}
\arguments{
\item{model}{The type of the model that the project is tied to}
}
\description{
Lists all the projects currently in the PLETHEM user database, This also lists
It is ties to one of the models that PLETHEM supports, either PBPK or HT
}
