% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchModeScripts.R
\name{loadBatchChemicalData}
\alias{loadBatchChemicalData}
\title{load chemical csv that can be used with batch mode}
\usage{
loadBatchChemicalData(file_path = NULL)
}
\arguments{
\item{file_path}{full path to csv file}
}
\description{
Load the chemical CSV file needed for batch mode processing 
The CSV file has the following column headers:
Name, 
Molecular Weight, 
LogP, 
LogKow,
Vapor Pressure,
Water Solubility,
Km,
Vmaxc,
Clintc,
IVIVE Type,
IVIVE Val1,
IVIVE Val2,
}
