% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchModeScripts.R
\name{loadBatchExposureData}
\alias{loadBatchExposureData}
\title{load exposure csv that can be used with batch mode}
\usage{
loadBatchExposureData(file_path = NULL)
}
\arguments{
\item{file_path}{full path to csv file}
}
\description{
Load the exposure CSV file needed for batch mode processing 
The CSV file has the following column headers:
Exposure Type, 
Exposure value 1,  
Exposure Value 2, 
Exposure Value 3,
Exposure Value 4,
Exposure Value 5
The values represent different parameters based on the type of exposure selected
Not all values are needed for each exposure type
For Bolus oral
1. Bolus Dose in mg/kg
2. number of bolus doses in a day
3. Logical TRUE/FALSE for repeating bolus oral dose
}
