% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{loadProject}
\alias{loadProject}
\title{Load the project with the given name}
\usage{
loadProject(file_path = "")
}
\arguments{
\item{file_path}{path to the project file}
}
\description{
Loads the project data in the PLETHEM project database.
}
