% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTRA.R
\name{loadTRA}
\alias{loadTRA}
\title{run TRA gadget and save files as needed.}
\usage{
loadTRA()
}
\description{
Run TRA gadget and save CSV files needed to export TRA exposures for the PBPK model
}
\examples{
if(interactive()){
loadTRA()
}
}
