% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{mainDbSelect}
\alias{mainDbSelect}
\title{Runs all select queries to the main database.}
\usage{
mainDbSelect(query, db_path = "")
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{The location of the project database. This defaults to database/plethemdb.sqlite and is not expected to change
This function will not be called by the user directly}
}
\description{
The function runs the select queries issued to the main db and returns the dataframe
}
