% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{mainDbUpdate}
\alias{mainDbUpdate}
\title{Runs all update queries to the main database.}
\usage{
mainDbUpdate(query, db_path = "database/plethemdb.sqlite")
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{The location of the main databse. This defualts to database/plethemdb.sqlite and is not expected to change
This function will not be called by the user directly}
}
\description{
The function runs the update queries issued to the main db
}
