% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcDataExample.R
\docType{data}
\name{mcDataExample}
\alias{mcDataExample}
\title{Monte Carlo Example Data Set For Testing The runReverseDosimetry Function}
\format{
An object of class \code{data.frame} with 1000 rows and 25 columns.
}
\source{
\href{https://www.scitovation.com/plethem/}{PLETHEM}
}
\usage{
data(mcDataExample)
}
\description{
This dataset was produced from the rapidPBPK PLETHEM module and is intended
to be used solely for the automated testing of the runReverseDosimetry
function.
}
\examples{
data(mcDataExample)
}
\references{
\href{https://www.scitovation.com/plethem/}{PLETHEM}
}
\keyword{Carlo}
\keyword{Data}
\keyword{Monte}
