% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEditVariabilityModule.R
\name{newEditVariabilityUI}
\alias{newEditVariabilityUI}
\title{UI function for defining variability and uncertainty datasets}
\usage{
newEditVariabilityUI(namespace)
}
\arguments{
\item{namespace}{namespace for the module when it is called form the PBPK UI}
}
\description{
UI function for defining variability and uncertainty datasets in the rapidPBPK model. This should not be called by the user
}
