% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{newProject}
\alias{newProject}
\title{Start a new PLETHEM project.}
\usage{
newProject(name = "new_project", type = "PBPK", model = "rapidPBPK",
  mode = "FD")
}
\arguments{
\item{name}{Name of the project}

\item{type}{The type of the model that the project is tied to}

\item{model}{The model to be used for the project}

\item{mode}{Either Forward Dosimetry(FD) or Monte Carlo(MC) mode. Only valid for PBPK type models}
}
\description{
A project consists of chemicals, organisms and datasets.
It is ties to one of the models that PLETHEM supports, either PBPK or HT
}
