% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProjectGadget.R
\name{newProjectGadget}
\alias{newProjectGadget}
\title{Command line function launching the new project gadget}
\usage{
newProjectGadget()
}
\description{
This launches the new project gadget for launching a new project in PLETHEM. With RStudio, this function is available as an addin
}
\examples{
if(interactive()){
newProjectGadget()
}
}
