% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseExposures.R
\name{parseConsExpoFile}
\alias{parseConsExpoFile}
\title{Parse uploaded file for consexpo}
\usage{
parseConsExpoFile(path)
}
\arguments{
\item{path}{Path to Excel File}
}
\description{
Parses the uploaded ConsExpo exposure SpreadSheet to extract all the exposure names and values
This function will not be called directly by the user
}
