% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseExposures.R
\name{parseTRAFile}
\alias{parseTRAFile}
\title{Parse uploaed file for Consumer TRA}
\usage{
parseTRAFile(path)
}
\arguments{
\item{path}{Path to Excel File}
}
\description{
Parses the uploaded Consumer TRA exposure SpreadSheet to extract all the exposure names and values
This function will not be called directly by the user
}
