% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{perform_HTIVIVE}
\alias{perform_HTIVIVE}
\title{Main function to run HT-Reverse Dosimetry from command line}
\usage{
perform_HTIVIVE(chemical, physiological, type, POD, hepatic, renal, plasma)
}
\arguments{
\item{chemical}{list of chemicals}

\item{physiological}{List of lists for physiological parameter}

\item{type}{list of type of exposure to be assumed}

\item{POD}{list of invitro point of depature in um}

\item{hepatic}{list of lists for hepatic clearance}

\item{renal}{list of TRUE/FALSE if renal clearance should be assumed}

\item{plasma}{list of plasma clearance values}
}
\description{
Main function to run HT-Reverse Dosimetry from command line
}
