% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performIVIVEModule.R
\name{performIVIVE}
\alias{performIVIVE}
\title{server side function for performing IVIVE for a chemical}
\usage{
performIVIVE(input, output, session, km)
}
\arguments{
\item{input}{input object for the UI}

\item{output}{input object to the UI}

\item{session}{session object for the module}

\item{km}{km for the chemical}
}
\description{
The server side function for running the IVIVE module from within the PBPK UI
}
