% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performIVIVEModule.R
\name{performIVIVEUI}
\alias{performIVIVEUI}
\title{Shiny module that is called when perform IVIVE button is clicked on the chemical tab of a PBPK model}
\usage{
performIVIVEUI(namespace)
}
\description{
Shiny module that is called when perform IVIVE button is clicked on the chemical tab of a PBPK model
}
