% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCAfunctions.R
\name{performPlethemNCA}
\alias{performPlethemNCA}
\title{Functions for performing NCA and returning the data}
\usage{
performPlethemNCA(result, var_names, mode = "FD")
}
\arguments{
\item{result}{Result from the forward dosimetry or individual montecarlo run}

\item{mode}{Forward Dosimetry or Montecarlo}
}
\description{
Calculates  the The user will never need to call this function
}
\details{
Calculate AUC, Cmax and Time to Cmax from the results of PBPK model run
}
