\name{plethem-package}
\alias{plethem-package}
\alias{plethem}
\docType{package}
\title{
\packageTitle{plethem}
}
\description{
  \packageDescription{plethem}
}

\author{

Jeremy Fitzpatrick, 
Kevin Bronson, 
C. Eric Hack, 
Patrick McMullen, 
and Salil Pendse

}
\references{
\href{https://doi.org/10.1016/j.comtox.2019.100115}{Pendse, S.N.; Efremenko, A.; Hack, C.E.; Moreau, M.; Mallick, P.; Dzierlenga, M.; Nicolas, C.I.; Yoon, M.; Clewell, H.J.; McMullen, P.D. Population Life-course Exposure to Health Effects Model (PLETHEM): An R package for PBPK modeling. Computational Toxicology. 2020 13:1-10}
}


\examples{
  ## run the interactive interface for foward dosimetry
  interactivePBPK(name='rapidPBPK')
}
