% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{preprocessUIData}
\alias{preprocessUIData}
\title{preprocess value list from UI}
\usage{
preprocessUIData(val)
}
\arguments{
\item{val}{list containing data for each row from the UI}
}
\value{
list of cleranace values for point of departure values and type of reverse dosimetry
}
\description{
The function converts the data from the UI to standard units of liters/h for clearances
and mg/L for concentrations. This function is not available to the end user as it relied on UI names
}
