% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{projectDbSelect}
\alias{projectDbSelect}
\title{Runs all select queries to the project database.}
\usage{
projectDbSelect(query, db_path = "database/project.sqlite")
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{The location of the project databse. This defualts to database/project.sqlite and is not expected to change
This function will not be called by the user directly}
}
\description{
The function runs the select queries issued to the project db and returns the dataframe
}
