% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{projectDbUpdate}
\alias{projectDbUpdate}
\title{Runs all update queries to the project database.}
\usage{
projectDbUpdate(query, db_path = "database/project.sqlite")
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{The location of the project databse. This defualts to database/project.sqlite and is not expected to change
This function will not be called by the user directly}
}
\description{
The function runs the update queries issued to the project db
}
