% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{projectReadTable}
\alias{projectReadTable}
\title{Read all the contents of a table}
\usage{
projectReadTable(tble_name, db_path = "database/project.sqlite")
}
\arguments{
\item{tble_name}{Name of the table to save. NOTE SQLITE table names are not case sensitive}

\item{db_path}{The location of the project database. This defaults to database/project.sqlite and is not expected to change}
}
\value{
table as a dataframe
}
\description{
The functions reads all the contents of the given table.
It is used internally to save project data
}
