% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{projectWriteTable}
\alias{projectWriteTable}
\title{Write the dataframe to the table}
\usage{
projectWriteTable(data, tble_name, db_path = "")
}
\arguments{
\item{data}{Data frame containing the data to save to the table. This will overwrite any existing data}

\item{tble_name}{Name of the table to write the data to. NOTE SQLITE table names are not case sensitive}

\item{db_path}{The location of the project database. This defaults to database/project.sqlite and is not expected to change}
}
\value{
None
}
\description{
The functions writes the datafarme to the sqlite table. It will overwrite any data
It is used internally to load project data
}
