% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QsarModels.R
\name{qsarModeltwo}
\alias{qsarModeltwo}
\title{Calculate Partition Coefficient using the default QSAR model}
\usage{
qsarModeltwo(chem_params, selected_org, tissue_list)
}
\arguments{
\item{chem_params}{A named list of parameters needed to run the model}

\item{selected_org}{Either "human" or "rat"}

\item{tissue_list}{List of tissues for which partition coefficients have to be calculated}
}
\value{
Named list of partition coefficients, one for each tissue in the tissue list
}
\description{
Calculates the partition coefficient using the default QSAR model. This QSAR model is based on the one described by Jongneelan et al as a part of the IndusChemFate model
}
