% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChemicals.R
\name{readBatchChemicalFile}
\alias{readBatchChemicalFile}
\title{Read and parse the chemical batch file}
\usage{
readBatchChemicalFile(file_path)
}
\arguments{
\item{file_path}{path to the batch chemical csv file}
}
\description{
Function to read the chemical batch file. The batch file description can be found in the user guide.
}
