% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChemicals.R
\name{readOperaPredictions}
\alias{readOperaPredictions}
\title{Read and parse the predictions from OPERA}
\usage{
readOperaPredictions(file_path)
}
\arguments{
\item{file_path}{path to OPERA predictions}
}
\description{
Function to read the OPERA predictions. The function removes predictions made by OPERA that are not used by PLETHEM.
}
