% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchModeScripts.R
\name{runBatchMode}
\alias{runBatchMode}
\title{Run batch mode}
\usage{
runBatchMode(chemicals = NULL, exposures = NULL, load_files = F,
  model = "rapidPBPK", organism = "human")
}
\arguments{
\item{chemicals}{if load_files = True, string containing the path to chemical csv file
or a data frame as generated by \code{\link{loadBatchChemicalData}}}

\item{exposures}{if load_files = True, string containing the path to the exposure csv file
else a data frame as generated by \code{\link{loadBatchExposureData}}}

\item{load_files}{if TRUE, expects chemicals and exposures parameters to be strings containing full path to those files
else expects data frames}

\item{model}{Model against which all simulations are to be run 
defaults to rapidPBPK}

\item{organism}{Organism for which the model should be parameterised for all runs
defaults to human}

\item{save_to_file.}{If true promts the user to select a path to save a excel file with all the result}
}
\value{
If save_to_file is FALSE, returns a list of dataframes, each contaning simulation results, indexed by compound name.
}
\description{
Runs simulations in batch mode.
}
