% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModels.R
\name{runFDPBPK}
\alias{runFDPBPK}
\title{Run the PBPK models in forward dosimetry mode}
\usage{
runFDPBPK(initial_values, model = "rapidPBPK")
}
\arguments{
\item{initial_values}{A list containing initial values needed to run the model}

\item{model}{The name of the PBPK model to simulate}
}
\description{
Run the pbpk model in forward dosimetry mode.
This function is common across all PBPK models.
This function can be used from the console if all the inputs are provided
}
