% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE.R
\name{runPlthemHTIVIVE}
\alias{runPlthemHTIVIVE}
\title{High Throughput reverse dosimetry calculation using UI}
\usage{
runPlthemHTIVIVE(vals)
}
\arguments{
\item{vals}{values passed from the PLETHEM UI}
}
\value{
List of oral equivalent dose, steady state plasma concentration and steady state
for each name in vals
}
\description{
Main function called from PLETHEM UI to run HT- Reverse Dosimetry
}
