% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Route2RouteExtrapolationModule.R
\name{runRoute2RouteUI}
\alias{runRoute2RouteUI}
\title{UI for performing route to route extrapolation}
\usage{
runRoute2RouteUI(namespace, simid, model)
}
\arguments{
\item{namespace}{namespace for the current instance of the module.}

\item{simid}{The id of the simulation to use for performing route to route extrapolation}

\item{model}{String representing the model that is running the simulation}
}
\description{
Module for running route to route extrapolation. This function is never called by the user.
}
