% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParameterSetModule.R
\name{saveAsParameterSet}
\alias{saveAsParameterSet}
\title{server side function for saving a new physiological, chemical or exposure set to the project database}
\usage{
saveAsParameterSet(input, output, session, set_type, main_input, name_df)
}
\description{
server side function for saving a new physiological, chemical or exposure set to the project database
}
