% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParameterSetModule.R
\name{saveAsParameterSetUI}
\alias{saveAsParameterSetUI}
\title{UI for saving a new physiological, chemical or exposure set to the project database}
\usage{
saveAsParameterSetUI(namespace, set_type)
}
\arguments{
\item{namespace}{namespace for the UI}

\item{set_type}{type of parameter set to save}
}
\description{
UI for saving parameter sets. This function should not be called by the user
}
