% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{saveProject}
\alias{saveProject}
\title{Save the current project to a location}
\usage{
saveProject()
}
\description{
Save the current PBPK or HTIVIVE project the user is working on
This cannot be used to save exposure or IVIVE gadget data.
This function should not be called directly from the console. It will be called by the app on exit
}
