% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRestoreParameterSetModule.R
\name{saveRestoreParameterSet}
\alias{saveRestoreParameterSet}
\title{Server for the restore/save dialog}
\usage{
saveRestoreParameterSet(input, output, session, UI_values, set_values,
  param_names, type)
}
\description{
Server for the save restore dialog box. This function should not be called by the user.
}
