% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRestoreParameterSetModule.R
\name{saveRestoreParameterSetUI}
\alias{saveRestoreParameterSetUI}
\title{UI for the restore/save dialog}
\usage{
saveRestoreParameterSetUI(namespace)
}
\description{
UI for the save restore dialog box. This function should not be called by the user.
}
