% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModels.R
\name{simulateRapidPBPKModel}
\alias{simulateRapidPBPKModel}
\title{Simulate the rapidPBPK model}
\usage{
simulateRapidPBPKModel(parameters, initial_states)
}
\arguments{
\item{parameters}{Parameter values for static parameters in the model.}

\item{init_states}{Initial values for all state variables in the model.}

\item{event_times}{Times for all events in the model.}

\item{times}{Times at which values need to be output.}

\item{outputs}{A list of of variables which are output by model.}
}
\value{
modelOutput DataFrame containing timecourse for all the state variables and outputs from the model
}
\description{
Simulates the rapidPBPK model within PLETHEM and returns the timecourse for all state variables.
This function can be called from the console if all inputs are provided
}
