% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUI.R
\name{updateUIInputs}
\alias{updateUIInputs}
\title{Update Inputs for PLETHEM UI}
\usage{
updateUIInputs(session, param_df)
}
\arguments{
\item{session}{The Shiny session in which all the inputs have to be updated}

\item{param_df}{The parameter dataframe. The parameter dataframe has the following columns
Name - The name of the parameter
Var - The variable in the UI representing this parameter
ParamType - The type of input, either Numeric, Radio, Checkbox, Select, Tabset. Used to identify which update function to call
Val - The value to be update with. If the type is numeric, the value is coerced to be a number.}
}
\description{
This a common function used to update the inputs for any PLETHEM User Interface. It is used by the apps to
provide a common pathways for all UI updates to happen
}
