% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{userDbSelect}
\alias{userDbSelect}
\title{Runs all select queries to the user database.}
\usage{
userDbSelect(query)
}
\arguments{
\item{query}{A valid SQL Query}
}
\description{
The function runs the select queries issued to the user db and returns the dataframe
the path to user database is stored in main plethem database and is selected from there
}
