% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{userDbUpdate}
\alias{userDbUpdate}
\title{Runs all update queries to the user database.}
\usage{
userDbUpdate(query)
}
\arguments{
\item{query}{A valid SQL Query}

\item{db_path}{The location of the user databse. This defualts to database/user.sqlite and is not expected to change
This function will not be called by the user directly}
}
\description{
The function runs the update queries issued to the user db
}
