% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{addParameters}
\alias{addParameters}
\title{Pipe to add concentration-timecourse data to the report}
\usage{
addParameters(HESI_doc, chemical, exposure, physiological)
}
\arguments{
\item{report_doc}{an officer::read_docx object}

\item{context}{a data.frame of concentrationXtime data, as returned by concData()}

\item{conc_units}{a string denoting the units of concentration}
}
\value{
report_doc, the same officer::read_docx object
}
\description{
add user defined parameter tables
}
