% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HESIreport.R
\name{createHESIgraphs}
\alias{createHESIgraphs}
\title{Gets the metabolism data. Should not be used by directly by the user}
\usage{
createHESIgraphs()
}
\arguments{
\item{admeid}{The id for ADME set. The admeid is used to obtain information about the other sets.}
}
\value{
List containing the metabolism values needed to run PBPK model or
display simulation information
}
\description{
The function returns the relavent metabolism data if the simulation contains
data from the metabolism set
}
