% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{getFileFolderPath}
\alias{getFileFolderPath}
\title{Show dialogs to select files or folders}
\usage{
getFileFolderPath(type = "dir", caption = "", extension = "", new_flag = FALSE
}
\arguments{
\item{type}{Type of document to get a path for. dir for directory or file for file}

\item{caption}{Caption to display for choose file/directory modal}

\item{extension}{extensions to display for selecting file type}

\item{new_flag}{Logical value for the "new" parameter in file.choose. Only used on MacOS}
}
\value{
path to the selected file or directory
}
\description{
The function shows the dialog to select files or folders. The functions change depending on the OS in which 
RStudio is running. It is only called internally and should not be run by the user.
}
