% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpkUtils.R
\name{getMetabData}
\alias{getMetabData}
\title{Gets the metabolism data. Should not be used by directly by the user}
\usage{
getMetabData(metabid, physioid, chemid, model = "rapidPBPK")
}
\arguments{
\item{metabid}{The meabolism id to get information from the metabolism tables}

\item{physioid}{Physiological id to get the age from the physiological table}
}
\value{
List containing the metabolism values needed to run PBPK model or
display simulation information
}
\description{
The function returns the relavent metabolism data if the simulation contains
data from the metabolism set
}
