% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{getProjectChemicalList}
\alias{getProjectChemicalList}
\title{get the list of chemicals currently in the project database}
\usage{
getProjectChemicalList()
}
\description{
Gets the chemicals that are currently a part of the project. This list comes from the chemical table in the
  project database. This function returns the chemical properties needed by all the models
  within PLETHEM. Chemical information specific to the model currently used is requested through a different function. This
  function can be used directly by the user to get a list of chemicals in the current project.
}
