% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE_InputModule.R
\name{HT_IVIVE}
\alias{HT_IVIVE}
\title{server function of high throughput dosimetry
This function is needed internally by the package and depends on unexported functions from it.}
\usage{
HT_IVIVE(input, output, session, vals = "", type = "", chem_list = list(),
  idx = 0, row_selected = 0)
}
\description{
server function of high throughput dosimetry
This function is needed internally by the package and depends on unexported functions from it.
}
