% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_IVIVE_InputModule.R
\name{HT_IVIVEUI}
\alias{HT_IVIVEUI}
\title{Module for editing high throughput reverse dosimetry functions}
\usage{
HT_IVIVEUI(namespace = "", set_list = NULL)
}
\arguments{
\item{namespace}{namespace for the module. This is unique and decided by the project server function}

\item{set_list}{A list of inputs for the dropdown menus.}
}
\description{
The UI for defining HT-IVIVE parameters in the HT-IVIVE project.
It is called by the HT-IVIVE server script when a new row is added or existing row is edited. It is never called directly by the user.
}
