% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httkOralEqDoseGadget.R
\name{httkCalcOralEqDose}
\alias{httkCalcOralEqDose}
\title{Function that runs the httk oral equivalent dose gadget.}
\usage{
httkCalcOralEqDose()
}
\description{
Launches the HTTK oral equivalent dose gadget. It servers as a UI wrapper around HTTK's functions to calculate oral equivalent dose.
}
\examples{
\dontrun{
httkCalcOralEqDose()
}
}
