% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllExposureModule.R
\name{importAllExposureDataUI}
\alias{importAllExposureDataUI}
\title{UI for importing all (Batch Exposure, TRA, SEEM, and SHEDS) data module.}
\usage{
importAllExposureDataUI(namespace)
}
\arguments{
\item{namespace}{namespace for the module UI.}
}
\description{
This function is called by the pbpk model to import all (Batch Exposure, TRA, SEEM, and SHEDS) data module estimates. Never called by the user
}
