% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHTTKDataModule.R
\name{importHTTKDataUI}
\alias{importHTTKDataUI}
\title{UI for importing data from HTTK}
\usage{
importHTTKDataUI(namespace)
}
\description{
UI function for the shiny module called by the rapidPBPK Exposure UI.
This will never be called directly by the user.
}
