% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importParameterSetModule.R
\name{importParameterSet}
\alias{importParameterSet}
\title{Server for import parameter module}
\usage{
importParameterSet(input, output, session, set_type, module_source = "PBPK")
}
\arguments{
\item{input}{input object from the UI}

\item{output}{output object for the UI}

\item{session}{session object for the server}

\item{set_type}{type of data to be imported}
}
\description{
server function for importing parameter sets from user databases. This function should never be called by the user
}
