% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importParameterSetModule.R
\name{importParameterSetUI}
\alias{importParameterSetUI}
\title{UI for importing parameter sets}
\usage{
importParameterSetUI(namespace, set_type)
}
\arguments{
\item{namespace}{namespace for this module}

\item{set_type}{type of data to be imported; physiological, chemical or exposure}
}
\description{
UI for importing parameters from user or main databases
}
