% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSHEDSDataModule.R
\name{importShedsData}
\alias{importShedsData}
\title{Server function for importing SHEDS data}
\usage{
importShedsData(input, output, session, path, expo_name_df)
}
\arguments{
\item{input}{input object for the UI}

\item{output}{input object to the UI}

\item{session}{session object for the module}

\item{path}{path where SHEDS results are stored}

\item{expo_name_df}{dataframe containing variable names for exposure values}
}
\description{
Server function for importing SHEDS exposures estimates into PLETHEM. The function should not be called by the user
}
