% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactivePBPK}
\alias{interactivePBPK}
\title{Launch the interactive PBPK workflow for the given model}
\usage{
interactivePBPK(name = "rapidPBPK")
}
\arguments{
\item{name}{Name of the model. "rapidPBPK" or "fishPBPK". Defaults to rapidPBPK.}
}
\description{
Used to launch the PBPK workflow for the given model. This interface can be used to launch either the rapidPBPK model, The HTTK model or the fishPBPK model.
}
\examples{
if(interactive()){
interactivePBPK("rapidPBPK")
}
}
