% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{interactiveReverseDosimetry}
\alias{interactiveReverseDosimetry}
\title{Launch Reverse Dosimetry Interface}
\usage{
interactiveReverseDosimetry()
}
\description{
Used to launch the reverse dosimetry UI. This UI allows the user to perform reverse dosimetry if they have already run Monte Carlo Anlaysis outside of PLETHEM.
}
\examples{
if(interactive()){
interactiveReverseDosimetry()
}
}
\seealso{
\code{\link{interactivePBPK}} for running reverse dosimetry using the rapidPBPK model in PLETHEM.
}
