% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEditVariabilityModule.R
\name{newEditVariability}
\alias{newEditVariability}
\title{Server side function for variability}
\usage{
newEditVariability(input, output, session, set_type, ops_type, var_params_list,
  set_id = 0)
}
\description{
Server side function for variability
}
