% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProjectGadget.R
\name{newProjectGadget}
\alias{newProjectGadget}
\title{Command line function launching the new project gadget}
\usage{
newProjectGadget()
}
\description{
This launches the new project gadget for launching a new project in PLETHEM. If using Rstudio Addins, the addin also calls this function.
}
