% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performIVIVEModule.R
\name{performIVIVE}
\alias{performIVIVE}
\title{server side function for performing IVIVE for a chemical}
\usage{
performIVIVE(input, output, session, km)
}
\description{
server side function for performing IVIVE for a chemical
}
