% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performIVIVEModule.R
\name{performIVIVEUI}
\alias{performIVIVEUI}
\title{Shiny module that is called when perform IVIVE button is clicked on the chemical tab of a PBPK model}
\usage{
performIVIVEUI(namespace)
}
\arguments{
\item{namespace}{namespace for the UI}
}
\description{
UI for performing IVIVE from within the PBPK UI. This function should not be called by the user
}
