% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performReverseDosimetry.R
\name{runReverseDosimetry}
\alias{runReverseDosimetry}
\title{Estimate exposure from montecarlo results and biomonitoring data}
\usage{
runReverseDosimetry(
  mcData,
  biomData,
  exposureBounds,
  percentiles = c(5, 50, 95, 99)
)
}
\arguments{
\item{mcData}{M by N data frame where M is the the individual exposures at which the PBPK model is run and N is the number of monte carlo runs at each exposure. It is recmmonded that M is between 25 and 40 and N is greater than 1000}

\item{biomData}{List consisting of biomonitering data. It is recommeded that atleast 1000 biomonitering values be provided to ensure accuracy for results}

\item{exposureBounds}{upper and lower bounds of the exposure for which the mcData was run}

\item{percentiles}{vector of percentiles for which exposure needs to be estimated. By default returns the 5th, 50th, 95th and 99th exposure estimate}
}
\value{
List of values related to reverse dosimetry
\describe{
\item{cdf}{Cumulative Distribution function of the exposure estimate}
\item{pdf}{Probility distribution function of the exposure estimate}
\item{percentiles}{Data frame of percentiles and exposure estimates for the percentile}

}
}
\description{
The function estimates exposures for the observed biomonitoring data using montecarlo simulation results over a large range of exposures.
The montecarlo results are obtained from a PBPK model. The biomonitoring results are obtained from a population level study.
The montecarlo results and biomonitoring data should have the same units and should be for the same physiological data source (eg: metabolite concentration in the urine).
}
\examples{
\dontrun{
runReverseDosimetry(mcData,biomdata,exposureBounds = c(0.01,1.4), percentiles = c(5,50,95))
}
}
