% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParameterSetModule.R
\name{saveAsParameterSet}
\alias{saveAsParameterSet}
\title{server side function for saving a new physiological, chemical or exposure set to the project database}
\usage{
saveAsParameterSet(
  input,
  output,
  session,
  set_type,
  main_input,
  name_df,
  other = NULL
)
}
\arguments{
\item{input}{input object for the UI}

\item{output}{input object to the UI}

\item{session}{session object for the module}

\item{set_type}{type of parameter set to save}

\item{main_input}{input from the pbpk UI}

\item{name_df}{variable names for parameters}

\item{other}{placeholder paramter for data needed for certain sets}
}
\description{
Server side function for running the save parameter module. This function should not be called by the user
}
