% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParameterSetModule.R
\name{saveAsParameterSetUI}
\alias{saveAsParameterSetUI}
\title{UI for saving a new physiological, chemical or exposure set to the project database}
\usage{
saveAsParameterSetUI(namespace, set_type)
}
\description{
UI for saving a new physiological, chemical or exposure set to the project database
}
