% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVIVEScaling.R
\name{scale_enzymatic}
\alias{scale_enzymatic}
\title{Scale clerance by age for when enzymatic clearance is known at referance age.}
\usage{
scale_enzymatic(out_ages, tot_scaled_recomcl, cypDb, cypCl, gender)
}
\description{
Scale clerance by age for when enzymatic clearance is known at referance age.
}
