% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryHelper.R
\name{userDbUpdate}
\alias{userDbUpdate}
\title{Runs all update queries to the user database.}
\usage{
userDbUpdate(query)
}
\arguments{
\item{query}{A valid SQL Query}
}
\description{
The function runs the update queries issued to the user db
}
